# Slam/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Book Changer API
  version: "1.0.0"
  description: Документация API для сервиса обмена книгами.
servers:
  - url: https://book-changer.ru
paths:
...
  /user/profile:
    get:
      summary: Получить профиль пользователя
      requestBody:
        description: Токен авторизации
        required: true
        content:
          schema:
            $ref:
      responses:
        '200':
          description: Данные профиля
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
    put:
      summary: Обновить профиль пользователя
      requestBody:
        description: Обновленные данные профиля
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserProfileDto'
      responses:
        '200':
          description: Профиль обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileDto'
...
  /books:
    post:
      summary: Создать книгу
      requestBody:
        description: Данные книги
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDto'
      responses:
        '200':
          description: Книга создана
    get:
      summary: Получить список книг
      responses:
        '200':
          description: Список книг
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BookDto'
  /books/{id}:
    get:
      summary: Получить книгу по ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Данные книги
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookDto'
    put:
      summary: Обновить книгу
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        description: Обновленные данные книги
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookDto'
      responses:
        '200':
          description: Книга обновлена
    delete:
      summary: Удалить книгу
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Книга удалена
...
  /reviews:
    post:
      summary: Создать отзыв
      requestBody:
        description: Данные отзыва
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewDto'
      responses:
        '200':
          description: Отзыв создан
  /reviews/by-book/{bookId}:
    get:
      summary: Получить отзывы по книге
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов по книге
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewDto'
  /reviews/by-user/{userId}:
    get:
      summary: Получить отзывы пользователя
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Список отзывов пользователя
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewDto'
...
components:
  schemas:
...
    UserProfileDto:
      type: object
      properties:
        idUser:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        userName:
          type: string
        rating:
          type: integer
...
    BookDto:
      type: object
      properties:
        idBookLiterary:
          type: integer
        idAuthor:
          type: integer
        bookName:
          type: string
        note:
          type: string
...
    ReviewDto:
      type: object
      properties:
        idBookResponse:
          type: integer
        bookLiteraryId:
          type: integer
        userId:
          type: integer
        createAt:
          type: string
          format: date-time
        response:
          type: string
...